Solve issues with the getProfileToChoose request
